SRB2: stop GCC -O2 code crash in cht_init()
git-svn-id: svn://code.srb2.org/SRB2/trunk@3716 d18046d3-e9af-4449-a702-ac2b6c676ba2

+++ src/m_cheat.c
@@ -84,7 +84,7 @@ static byte cheat_xlate_table[256];
 void cht_Init(void)
 {
 	size_t i = 0;
-	byte pi = 0;
+	short pi = 0;
 	for (; i < 256; i++, pi++)
 	{
 		const int cc = SCRAMBLE(pi);
